+++ app-wearos/src/main/java/host/stjin/anonaddy/watchface/AnonAddyWatchFace.kt
@@ -4,7 +4,12 @@ import android.content.BroadcastReceiver
 import android.content.Context
 import android.content.Intent
 import android.content.IntentFilter
-import android.graphics.*
+import android.graphics.Canvas
+import android.graphics.Color
+import android.graphics.Paint
+import android.graphics.Rect
+import android.graphics.Typeface
+import android.os.Build
 import android.os.Bundle
 import android.os.Handler
 import android.os.Message
@@ -18,7 +23,8 @@ import android.view.WindowInsets
 import androidx.core.content.ContextCompat
 import host.stjin.anonaddy.R
 import java.lang.ref.WeakReference
-import java.util.*
+import java.util.Calendar
+import java.util.TimeZone
 import kotlin.random.Random
 
 /**
@@ -290,7 +296,11 @@ class AnonAddyWatchFace : CanvasWatchFaceService() {
             }
             mRegisteredTimeZoneReceiver = true
             val filter = IntentFilter(Intent.ACTION_TIMEZONE_CHANGED)
-            this@AnonAddyWatchFace.registerReceiver(mTimeZoneReceiver, filter)
+            if (Build.VERSION.SDK_INT >= Build.VERSION_CODES.TIRAMISU) {
+                this@AnonAddyWatchFace.registerReceiver(mTimeZoneReceiver, filter, Context.RECEIVER_EXPORTED)
+            } else {
+                this@AnonAddyWatchFace.registerReceiver(mTimeZoneReceiver, filter)
+            }
         }
 
         private fun unregisterReceiver() {