+++ app/src/main/java/host/stjin/anonaddy/ui/alias/AliasFragment.kt
@@ -1,7 +1,13 @@
 package host.stjin.anonaddy.ui.alias
 
 import android.app.Activity
-import android.content.*
+import android.content.BroadcastReceiver
+import android.content.ClipData
+import android.content.ClipboardManager
+import android.content.Context
+import android.content.Intent
+import android.content.IntentFilter
+import android.os.Build
 import android.os.Bundle
 import android.os.Handler
 import android.os.Looper
@@ -206,7 +212,12 @@ class AliasFragment : Fragment(), AddAliasBottomDialogFragment.AddAliasBottomDia
     override fun onResume() {
         super.onResume()
         setHasReachedTopOfNsv()
-        activity?.registerReceiver(mScrollUpBroadcastReceiver, IntentFilter("scroll_up"))
+        if (Build.VERSION.SDK_INT >= Build.VERSION_CODES.TIRAMISU) {
+            activity?.registerReceiver(mScrollUpBroadcastReceiver, IntentFilter("scroll_up"), Context.RECEIVER_EXPORTED)
+        } else {
+            activity?.registerReceiver(mScrollUpBroadcastReceiver, IntentFilter("scroll_up"))
+        }
+
     }
 
     private fun setOnClickListeners() {