+++ app/src/main/java/host/stjin/anonaddy/ui/recipients/RecipientsFragment.kt
@@ -4,6 +4,7 @@ import android.content.BroadcastReceiver
 import android.content.Context
 import android.content.Intent
 import android.content.IntentFilter
+import android.os.Build
 import android.os.Bundle
 import android.view.LayoutInflater
 import android.view.View
@@ -147,7 +148,11 @@ class RecipientsFragment : Fragment(),
     override fun onResume() {
         super.onResume()
         setHasReachedTopOfNsv()
-        activity?.registerReceiver(mScrollUpBroadcastReceiver, IntentFilter("scroll_up"))
+        if (Build.VERSION.SDK_INT >= Build.VERSION_CODES.TIRAMISU) {
+            activity?.registerReceiver(mScrollUpBroadcastReceiver, IntentFilter("scroll_up"), Context.RECEIVER_EXPORTED)
+        } else {
+            activity?.registerReceiver(mScrollUpBroadcastReceiver, IntentFilter("scroll_up"))
+        }
     }
 
     private fun setOnClickListener() {